Separate comment servlets in order to implement login servlets
