+++ portfolio/src/main/java/com/google/sps/comments/NewCommentServlet.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.sps.servlets;
+package com.google.sps.comments;
 
 import com.google.sps.comments.Comment;
 import com.google.gson.Gson;
